+++ packages/multiple-choice/configure/src/main.jsx
@@ -16,6 +16,9 @@ import Tooltip from '@material-ui/core/Tooltip';
 import Dialog from '@material-ui/core/Dialog';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import DialogActions from '@material-ui/core/DialogActions';
+import Typography from '@material-ui/core/Typography';
+import Info from '@material-ui/icons/Info';
+import { generateValidationMessage } from './utils';
 
 const { Panel, toggle, radio, dropdown } = settings;
 
@@ -74,6 +77,28 @@ const styles = (theme) => ({
       backgroundColor: color.disabled(),
     },
   },
+  flexContainer: {
+    display: 'flex',
+    alignItems: 'center',
+    marginBottom: '5px'
+  },
+  titleText: {
+    fontFamily: 'Cerebri Sans',
+    fontSize: '18px',
+    lineHeight: '19px',
+    color: '#495B8F',
+    marginRight: '5px'
+  },
+  tooltip: {
+    fontSize: '12px',
+    whiteSpace: 'pre',
+    maxWidth: '500px'
+  },
+  errorText: {
+    fontSize: '12px',
+    color: 'red',
+    paddingTop: '5px'
+  }
 });
 
 const Design = withStyles(styles)((props) => {
@@ -108,7 +133,7 @@ const Design = withStyles(styles)((props) => {
     settingsPanelDisabled,
     choicesLayout,
     spellCheck = {},
-    gridColumns,
+    gridColumns
   } = configuration || {};
   let { maxAnswerChoices } = configuration || {};
   const {
@@ -120,8 +145,10 @@ const Design = withStyles(styles)((props) => {
     promptEnabled,
     spellCheckEnabled,
     choices,
+    errors
   } = model || {};
 
+  const { choicesErrors, correctResponseError, answerChoicesError } = errors || {};
   const nrOfColumnsAvailable =
     choices && choices.length
       ? Array.from({ length: choices.length }, (_, i) => `${i + 1}`)
@@ -147,6 +174,8 @@ const Design = withStyles(styles)((props) => {
   if (limitChoicesNumber) {
     maxAnswerChoices = MAX_CHOICES;
   }
+
+  const validationMessage = generateValidationMessage(configuration);
 
   const Content = (
     <div>
@@ -181,7 +210,21 @@ const Design = withStyles(styles)((props) => {
           />
         </InputContainer>
       )}
-      {model.choices.map((choice, index) => (
+      <div className={classes.flexContainer}>
+        <Typography className={classes.titleText}>Choices</Typography>
+        <Tooltip
+          classes={{ tooltip: classes.tooltip }}
+          disableFocusListener
+          disableTouchListener
+          placement={'right'}
+          title={validationMessage}
+        >
+          <Info fontSize={'small'} color={'primary'}/>
+        </Tooltip>
+      </div>
+      {correctResponseError && <div className={classes.errorText}>{correctResponseError}</div>}
+      {answerChoicesError && <div className={classes.errorText}>{answerChoicesError}</div>}
+      {choices.map((choice, index) => (
         <div
           key={`choice-${index}`}
           className={classes.choiceConfigurationHolder}
@@ -202,6 +245,8 @@ const Design = withStyles(styles)((props) => {
             noLabels
             toolbarOpts={toolbarOpts}
             spellCheck={spellCheckEnabled}
+            error={choicesErrors && choicesErrors[choice.value] ? choicesErrors[choice.value] : null}
+            noCorrectAnswerError={correctResponseError}
           />
           {rationaleEnabled && (
             <InputContainer
@@ -259,12 +304,7 @@ const Design = withStyles(styles)((props) => {
           classes={{ tooltip: classes.tooltip }}
         >
           <Button
-            classes={{
-              root:
-                maxAnswerChoices && model.choices.length >= maxAnswerChoices
-                  ? classes.disableButton
-                  : undefined,
-            }}
+            classes={{ root: maxAnswerChoices && model.choices.length >= maxAnswerChoices && classes.disableButton }}
             className={classes.addButton}
             variant="contained"
             color="primary"
@@ -402,14 +442,8 @@ export class Main extends React.Component {
                 open: false,
               },
             });
-          },
-          onCancel: () =>
-            this.setState({
-              dialog: {
-                open: false,
-              },
-            }),
-        },
+          }
+        }
       });
     } else {
       model.choices.splice(index, 1);
@@ -427,40 +461,17 @@ export class Main extends React.Component {
       maxAnswerChoices = MAX_CHOICES;
     }
 
-    if (maxAnswerChoices && model.choices.length === maxAnswerChoices) {
-      this.setState({
-        dialog: {
-          open: true,
-          message: `There can't be more than ${maxAnswerChoices} choices.`,
-          onOk: () => {
-            this.setState({
-              dialog: {
-                open: false,
-              },
-            });
-          },
-          onCancel: () =>
-            this.setState({
-              dialog: {
-                open: false,
-              },
-            }),
-        },
-      });
-    } else {
-      model.choices.push({
-        label: '',
-        value: utils.firstAvailableIndex(
-          model.choices.map((c) => c.value),
-          0
-        ),
-        feedback: {
-          type: 'none',
-        },
-      });
-
-      this.props.onModelChanged(model);
+    if (maxAnswerChoices && model.choices.length >= maxAnswerChoices) {
+      return;
     }
+
+    model.choices.push({
+      label: '',
+      value: utils.firstAvailableIndex(model.choices.map((c) => c.value), 0),
+      feedback: { type: 'none' }
+    });
+
+    this.props.onModelChanged(model);
   };
 
   onChoiceChanged = (index, choice) => {
@@ -503,15 +514,18 @@ export class Main extends React.Component {
           model.choices = model.choices.map((c) => {
             if (correctFound) {
               c.correct = false;
+
               return c;
             }
 
             if (c.correct) {
               correctFound = true;
             }
+
             return c;
           });
         }
+
         onModelChanged(model, true);
         break;
       }