+++ packages/matrix/configure/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.20](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.19...@pie-element/matrix-configure@0.1.20) (2022-03-28)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
+## [0.1.19](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.18...@pie-element/matrix-configure@0.1.19) (2022-03-22)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
+## [0.1.18](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.17...@pie-element/matrix-configure@0.1.18) (2022-03-08)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
+## [0.1.17](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.16...@pie-element/matrix-configure@0.1.17) (2022-03-08)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
+## [0.1.16](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.15...@pie-element/matrix-configure@0.1.16) (2022-02-21)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
+## [0.1.15](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.14...@pie-element/matrix-configure@0.1.15) (2022-02-04)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
+## [0.1.14](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.13...@pie-element/matrix-configure@0.1.14) (2022-01-24)
+
+**Note:** Version bump only for package @pie-element/matrix-configure
+
+
+
+
+
 ## [0.1.13](https://github.com/pie-framework/pie-elements/compare/@pie-element/matrix-configure@0.1.12...@pie-element/matrix-configure@0.1.13) (2022-01-10)
 
 **Note:** Version bump only for package @pie-element/matrix-configure